Chop number 3
