--- backend/Img2mapAPI/Helper/__init__.py
@@ -1,5 +1,0 @@
-from . import FileHelper
-from . import ImageHelper
-from . import FileStorage
-from . import localrepository
-from . import GeorefTestFiles
