+++ backend/Img2mapAPI/Img2mapAPI.py
@@ -3,9 +3,6 @@ from fastapi.middleware.cors import CORSMiddleware
 from .routers import *
 
 
-
-
-
 app = FastAPI(
     title="Img2Map API",
     description="API for converting and georeferencing images",
@@ -27,18 +24,12 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-
-#todo: add routes to router
-# Default test route
+# Default route
 @router.get("/")
 async def root():
     #redirect to the documentation
     return {"message": "Welcome to the georeferencing API. Please refer to the documentation for more information. at /docs or /redoc"}
 
 #adding the routers to the app
-
-#Converter routes
 app.router.include_router(converters.router, prefix="/converter", tags=["converter"])
-
-#Goereferencing routes
 app.router.include_router(georefProject.router)
