--- backend/Img2mapAPI/Models/__init__.py
@@ -1,5 +1,0 @@
-from . import Point
-from . import PointList
-from . import Project
-
-__ALL__ = ["Point", "PointList", "Project"]
