+++ backend/Img2mapAPI/__init__.py
@@ -1,6 +1,6 @@
 from . import Img2mapAPI
 from . import routers
-from . import models
-from . import helper
+from . import utils
+from . import devOnly
 
-__all__ = ["Img2mapAPI", "routers", "models", "helper"]
+__all__ = ["Img2mapAPI", "routers", "utils", "devOnly"]
