+++ backend/Img2mapAPI/devOnly/GeorefTestFiles/testpoints/tpgaldopiggenturkartp3.json
