+++ backend/Img2mapAPI/utils/core/FileHelper.py
