+++ backend/Img2mapAPI/utils/core/ImageHelper.py
