+++ backend/Img2mapAPI/utils/core/georefHelper.py
@@ -4,39 +4,10 @@ from rasterio.transform import from_gcps #importing from_gcps to create a transf
 from rasterio.control import GroundControlPoint as GCP 
 from rasterio.crs import CRS #importing CRS for the default crs
 from ..models import PointList #importing the pointList model
+from .FileHelper import getUniqeFileName, removeFile #importing the getUniqeFileName and removeFile functions from FileHelper
 
 #mapbox projection crs for web mercator: EPSG:3857
 defaultCrs = 'EPSG:4326'
-
-#checking if /tmp folder exists
-def getTmpFolderPath():
-    if not os.path.exists('./temp'):
-        os.makedirs('./temp')
-    return './temp'
-
-def getUniqeFileName(suffix, length=8):
-    #create a random file name between 8 cha, xxxx-xxxx.
-    import random
-    import string
-    #check if the file name is unique in the temp folder
-    i = 0
-    while True:
-        #create a random file name
-        randomName = ''.join(random.choices(string.ascii_lowercase + string.digits, k=length))
-        #check if the file name exists
-        if not os.path.isfile(f"./temp/{randomName}.{suffix}"):
-            return f"/{randomName}.{suffix}"
-        #raise an exception if max tries is reached
-        if i == 100:
-            raise Exception("Could not create a unique file name")
-        i += 1
-
-def removeFile(filePath):
-    #remove the file if it exists
-    if os.path.isfile(filePath):
-        os.remove(filePath)
-
-
 
 #function to create Rasterio GCPs from a list of points
 def createGcps(PointList : PointList):
