+++ backend/Img2mapAPI/utils/models/__init__.py
@@ -1,0 +1,5 @@
+from .point import Point
+from .pointList import PointList
+from .project import Project
+
+__ALL__ = ["Point", "PointList", "Project"]
