+++ backend/Img2mapAPI/utils/models/pointList.py
@@ -1,6 +1,6 @@
 from typing import List
 from pydantic import BaseModel
-from .Point import Point
+from .point import Point
 # model for the pointList
 
 class PointList(BaseModel):
