+++ backend/Img2mapAPI/utils/storage/__init__.py
@@ -1,0 +1,5 @@
+from . import filestorage
+from . import localrepository
+from . import storageHandler
+
+__ALL__ = ["filestorage", "localrepository", "storageHandler"]
