+++ backend/Img2mapAPI/utils/storage/filestorage/LocalFileStorage.py
@@ -1,21 +1,6 @@
-#Abstract class for file storage
-from abc import ABC, abstractmethod
 import os
 import tempfile
-
-class FileStorage(ABC):
-    @abstractmethod
-    def save(self, data: bytes, suffix: str)->str:
-        pass
-    @abstractmethod
-    def remove(self, path: str):
-        pass
-    @abstractmethod
-    def read(self, path: str)->bytes:
-        pass
-    @abstractmethod
-    def exists(self, path: str)->bool:
-        pass
+from .FileStorage import FileStorage
 
 #File storage using the local file system
 class LocalFileStorage(FileStorage):
